requirements.txt
torch==2.7.1
tsaug==0.2.1
scikit-learn==1.7.1
numpy==1.24.6
matplotlib==3.10.3
info-nce-pytorch==0.1.4
